requirement.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-autoslug==1.9.8
django-debug-toolbar==4.0.0
django-filter==23.1
django-pagination==1.0.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
page==0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2==2.9.5
PyJWT==2.6.0
python-slugify==8.0.1
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
validate-email==1.3